get venues,parentVenues,favVenues,venuedetailsitem,venueCategory datas from firebase firestore.Some fragments, viewModels, adapters updated

+++ app/src/main/java/com/imranmelikov/folt/presentation/discovery/VenueReviewAdapter.kt
@@ -76,10 +76,10 @@ class VenueReviewAdapter:RecyclerView.Adapter<VenueReviewAdapter.VenueReviewView
     private fun setRatingIcon(venueArraylist: Venue,ratingIcon: ImageView){
         when {
             // Rating less than 5,00
-            venueArraylist.venuePopularity.rating < 5.00 ->{
+            venueArraylist.venuePopularity.rating.toDouble() < 5.00 ->{
                 ratingIcon.setImageResource(R.drawable.emoticon_sad)
             }
-            venueArraylist.venuePopularity.rating <9.00 -> {
+            venueArraylist.venuePopularity.rating.toDouble() <9.00 -> {
                 //Rating less than 9.00
                 ratingIcon.setImageResource(R.drawable.emoticon_happy)
             }
